Split proxy code into a new file and add documentation.
